refactor: move authentication functions to authService and add utilService
